refactor: CLI main.rs を arguments / filter_mapping / depth_resolver に分割（振る舞い不変） (#159)
